Run flake8 and coverage with explicit config files.
With these additions, we are no longer checking folders containing intermediate
files and we don't include in coverage files that we don't intend to run in
the test suite by design.  Also, we are now running flake8 on the files in the
subpackages as should be the case.  Had to do some cleaning to get those
passing.